+++ lib/ch/row_binary.ex
@@ -196,7 +196,7 @@ defmodule Ch.RowBinary do
     # assuming it can be sent as text and not "native" binary JSON
     # i.e. assumes `settings: [input_format_binary_read_json_as_string: 1]`
     # TODO
-    encode(:string, Jason.encode_to_iodata!(json))
+    encode(:string, JSON.encode_to_iodata!(json))
   end
 
   def encode({:fixed_string, size}, str) when byte_size(str) == size do
@@ -880,7 +880,7 @@ defmodule Ch.RowBinary do
            rows,
            types
          ) do
-      decode_rows(types_rest, bin, [Jason.decode!(s) | row], rows, types)
+      decode_rows(types_rest, bin, [JSON.decode!(s) | row], rows, types)
     end
   end