+++ test/ch/connect_test.exs
@@ -1,5 +1,5 @@
 defmodule Ch.ConnectTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
   import ExUnit.CaptureLog
 
   @tag :slow