+++ test/ch/query_string_test.exs
@@ -1,7 +1,5 @@
 defmodule Ch.QueryStringTest do
-  use ExUnit.Case,
-    async: true,
-    parameterize: [%{query_options: []}, %{query_options: [multipart: true]}]
+  use ExUnit.Case, async: true
 
   setup ctx do
     {:ok, query_options: ctx[:query_options] || []}