+++ test/ch/query_test.exs
@@ -1,8 +1,5 @@
 defmodule Ch.QueryTest do
-  use ExUnit.Case,
-    async: true,
-    parameterize: [%{query_options: []}, %{query_options: [multipart: true]}]
-
+  use ExUnit.Case, async: true
   alias Ch.Query
 
   setup ctx do