+++ test/test_helper.exs
@@ -1,41 +1,21 @@
-clickhouse_available? =
-  case :httpc.request(:get, {~c"http://localhost:8123/ping", []}, [], []) do
-    {:ok, {{_version, _status = 200, _reason}, _headers, ~c"Ok.\n"}} ->
-      true
+# check if clickhouse is available
+case Help.http("http://localhost:8123/ping") do
+  {:ok, 200, _headers, "Ok.\n"} ->
+    :ok
 
-    {:error, {:failed_connect, [{:to_address, _to_address}, {:inet, [:inet], :econnrefused}]}} ->
-      false
-  end
+  other ->
+    Mix.shell().error("""
+    ClickHouse is not detected at localhost:8123. Please start the local container with the following command:
 
-unless clickhouse_available? do
-  Mix.shell().error("""
-  ClickHouse is not detected at localhost:8123! Please start the local container with the following command:
+        docker compose up -d clickhouse
+    """)
 
-      docker compose up -d clickhouse
-  """)
-
-  System.halt(1)
+    System.halt(1)
 end
 
-Calendar.put_time_zone_database(Tz.TimeZoneDatabase)
-default_test_db = System.get_env("CH_DATABASE", "ch_elixir_test")
-Application.put_env(:ch, :database, default_test_db)
+%{rows: [[ch_version]]} = Help.ch("SELECT version()")
 
-Ch.Test.query(
-  "DROP DATABASE IF EXISTS {db:Identifier}",
-  %{"db" => default_test_db},
-  database: "default"
-)
-
-Ch.Test.query(
-  "CREATE DATABASE {db:Identifier}",
-  %{"db" => default_test_db},
-  database: "default"
-)
-
-%{rows: [[ch_version]]} = Ch.Test.query("SELECT version()")
-
-extra_exclude =
+exclude =
   if ch_version >= "25" do
     []
   else
@@ -43,4 +23,4 @@ extra_exclude =
     [:time, :variant, :json, :dynamic]
   end
 
-ExUnit.start(exclude: [:slow | extra_exclude])
+ExUnit.start(exclude: exclude)